refactor: projectApplyStatus를 독립된 feature로 분리
