+++ src/features/board/projectApplyStatus/type.ts
